requirements.txt
asgiref==3.8.1
Django==5.1.4
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
drf-nested-routers==0.94.1
factory_boy==3.3.1
Faker==33.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
